cloud-back/requirements.txt
aws-cdk-lib==2.143.0
constructs>=10.0.0,<11.0.0
boto3
botocore
Pillow==10.3.0
PyJWT==2.8.0